move tests in dedicated file
